@@ -7,16 +7,18 @@
         "build": "./node_modules/.bin/tsc",
         "watch": "./node_modules/.bin/tsc --watch",
         "exec": "./node_modules/.bin/ts-node",
-        "test": "node node_modules/.bin/mocha --require ts-node/register test/**/*.ts --color"
+        "test": "node_modules/.bin/mocha --require ts-node/register test/**/*.ts --color"
     },
     "author": "",
     "license": "ISC",
     "dependencies": {
-        "ts-node": "^4.1.0",
-        "typescript": "^2.6.2"
+        "ts-node": "^7.0.1",
+        "typescript": "^3.0.1"
     },
     "devDependencies": {
+        "@types/jasmine": "^2.8.8",
+        "@types/node": "^10.7.1",
         "chai": "^4.1.2",
-        "mocha": "^4.1.0"
+        "mocha": "^5.2.0"
     }
 }
